Move input logic to input section
