+++ src/Pages/Sections/Card/MatchCardBottom.tsx
@@ -1,10 +1,7 @@
 import { Button, Grid, Typography } from '@material-ui/core';
 import React, { ReactElement, useState } from 'react';
-import { useDispatch } from 'react-redux';
 
-import { IPrediction } from '../../../models/predictions.model';
 import { IFixtureWithScoreAndPredictions } from '../../../models/toto.models';
-import { changePrediction, postNewPrediction } from '../../../store/predictions/action-creators';
 import { hasBettingClosed } from '../../../utils/timeFunctions';
 import { getPrediction } from './card-functions';
 import MatchCardInput from './MatchCardInput';
@@ -15,39 +12,13 @@ interface IProps {
 }
 
 const MatchCardBottom: React.FC<IProps> = ({ fixtureWithPrediction, display }: IProps): ReactElement => {
-  const dispatch = useDispatch();
-  const [pGoalsHT, setpGoalsHT] = useState<number>(0);
-  const [pGoalsAT, setpGoalsAT] = useState<number>(0);
   const [showInput, setShowInput] = useState<boolean>(false);
 
   const {
-    id,
     eventTimeStamp,
     predictions: { pGoalsAwayTeam, pGoalsHomeTeam },
     status,
   } = fixtureWithPrediction;
-
-  const handleSubmit = () => {
-    const prediction: IPrediction = {
-      pGoalsHomeTeam: pGoalsHT,
-      pGoalsAwayTeam: pGoalsAT,
-      fixtureId: id,
-    };
-
-    Number.isInteger(pGoalsAwayTeam) || Number.isInteger(pGoalsHomeTeam)
-      ? dispatch(changePrediction(prediction))
-      : dispatch(postNewPrediction(prediction));
-
-    setShowInput(false);
-  };
-
-  const handleGoalsHomeTeam = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setpGoalsHT(Number(e.target.value));
-  };
-
-  const handleGoalsAwayTeam = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setpGoalsAT(Number(e.target.value));
-  };
 
   const setShowInputToFalse = () => setShowInput(false);
   const setShowInputToTrue = () => setShowInput(true);
@@ -67,13 +38,7 @@ const MatchCardBottom: React.FC<IProps> = ({ fixtureWithPrediction, display }: I
             </Button>
           </Grid>
         ) : status !== 'Match Finished' && !hasBettingClosed(eventTimeStamp) && display === 'private' && showInput ? (
-          <MatchCardInput
-            fixtureWithPrediction={fixtureWithPrediction}
-            hideInput={setShowInputToFalse}
-            handleGoalsAwayTeam={handleGoalsAwayTeam}
-            handleGoalsHomeTeam={handleGoalsHomeTeam}
-            handleSubmit={handleSubmit}
-          />
+          <MatchCardInput fixtureWithPrediction={fixtureWithPrediction} hideInput={setShowInputToFalse} />
         ) : null}
       </>
     </Grid>
